+++ src/charm.py
@@ -261,7 +261,7 @@ class WazuhServerCharm(CharmBaseWithState):
             self.state.cluster_key,
             opencti_token=self.state.opencti_token,
             opencti_url=self.state.opencti_url,
-            disable_vulnerability_detection=self.state.disable_vulnerability_detection,
+            enable_vulnerability_detection=self.state.enable_vulnerability_detection,
         )
         if any((updated_config, changed_password, changed_ossec_conf)):
             self._restart_service(container, WAZUH_SERVICE_NAME, force=True)