+++ tests/unit/test_charm.py
@@ -218,7 +218,7 @@ def test_reconcile_reaches_active_status_when_repository_and_password_configured
         cluster_key,
         opencti_url=opencti_url,
         opencti_token=opencti_token,
-        disable_vulnerability_detection=False,
+        enable_vulnerability_detection=True,
     )
     sync_filebeat_user_mock.assert_called_with(container, "user1", password)
     ensure_rsyslog_output_dir_mock.assert_called_with(container)
@@ -333,7 +333,7 @@ def test_reconcile_reaches_active_status_when_repository_and_password_not_config
         cluster_key,
         opencti_token=None,
         opencti_url=None,
-        disable_vulnerability_detection=False,
+        enable_vulnerability_detection=True,
     )
     get_version_mock.assert_called_with(container)
     assert harness.model.unit.status.name == ops.ActiveStatus().name