+++ docs/changelog.md
@@ -8,7 +8,7 @@ Each revision is versioned by the date of the revision.
 
 ## 2025-11-14
 
-- Added disable-vulnerability-detection config option.
+- Added enable-vulnerability-detection config option.
 
 ## 2025-10-15